refactor(cli): extract diagnostics command subsystem module (#170)
